+++ bin/collator/Cargo.toml
@@ -33,13 +33,13 @@ sp-api = { workspace = true, features = ["std"] }
 sp-block-builder = { workspace = true, features = ["std"] }
 sp-consensus-aura = { workspace = true, features = ["std"] }
 sp-core = { workspace = true, features = ["std"] }
+sp-genesis-builder = { workspace = true, features = ["std"] }
 sp-inherents = { workspace = true, features = ["std"] }
 sp-io = { workspace = true, features = ["std"] }
 sp-offchain = { workspace = true, features = ["std"] }
 sp-runtime = { workspace = true, features = ["std"] }
 sp-session = { workspace = true, features = ["std"] }
 sp-transaction-pool = { workspace = true, features = ["std"] }
-sp-genesis-builder = { workspace = true, features = ["std"] }
 
 sp-blockchain = { workspace = true }
 sp-consensus = { workspace = true }
