+++ bin/collator/src/parachain/chain_spec/astar.rs
@@ -18,20 +18,9 @@
 
 //! Astar chain specifications.
 
-use super::{get_from_seed, Extensions};
-use astar_primitives::oracle::CurrencyAmount;
-use astar_runtime::{
-    wasm_binary_unwrap, AccountId, AuraId, Balance, DappStakingConfig, EVMConfig, InflationConfig,
-    InflationParameters, OracleMembershipConfig, ParachainInfoConfig, Precompiles,
-    PriceAggregatorConfig, Signature, TierThreshold, ASTR,
-};
-use cumulus_primitives_core::ParaId;
+use super::Extensions;
+use astar_runtime::wasm_binary_unwrap;
 use sc_service::ChainType;
-use sp_core::{sr25519, Pair, Public};
-use sp_runtime::{
-    traits::{IdentifyAccount, Verify},
-    Perbill, Permill,
-};
 
 const PARA_ID: u32 = 2006;
 
@@ -39,21 +28,8 @@ const PARA_ID: u32 = 2006;
 pub type AstarChainSpec =
     sc_service::GenericChainSpec<astar_runtime::RuntimeGenesisConfig, Extensions>;
 
-/// Gen Astar chain specification for given parachain id.
+/// Get Astar chain specification.
 pub fn get_chain_spec() -> AstarChainSpec {
-    // Alice as default
-    let sudo_key = get_account_id_from_seed::<sr25519::Public>("Alice");
-    let endowned = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            1_000_000_000 * ASTR,
-        ),
-        (
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            1_000_000_000 * ASTR,
-        ),
-    ];
-
     let mut properties = serde_json::map::Map::new();
     properties.insert("tokenSymbol".into(), "ASTR".into());
     properties.insert("tokenDecimals".into(), 18.into());
@@ -70,154 +46,6 @@ pub fn get_chain_spec() -> AstarChainSpec {
     .with_id("astar")
     .with_chain_type(ChainType::Development)
     .with_properties(properties)
-    .with_genesis_config(make_genesis(
-        endowned.clone(),
-        sudo_key.clone(),
-        PARA_ID.into(),
-    ))
+    .with_genesis_config(astar_runtime::genesis_config::default_config(PARA_ID))
     .build()
-}
-
-fn session_keys(aura: AuraId) -> astar_runtime::SessionKeys {
-    astar_runtime::SessionKeys { aura }
-}
-
-/// Helper function to create RuntimeGenesisConfig.
-fn make_genesis(
-    balances: Vec<(AccountId, Balance)>,
-    root_key: AccountId,
-    parachain_id: ParaId,
-) -> serde_json::Value {
-    let authorities = vec![
-        (
-            get_account_id_from_seed::<sr25519::Public>("Alice"),
-            get_from_seed::<AuraId>("Alice"),
-        ),
-        (
-            get_account_id_from_seed::<sr25519::Public>("Bob"),
-            get_from_seed::<AuraId>("Bob"),
-        ),
-    ];
-
-    // This is supposed the be the simplest bytecode to revert without returning any data.
-    // We will pre-deploy it under all of our precompiles to ensure they can be called from
-    // within contracts.
-    // (PUSH1 0x00 PUSH1 0x00 REVERT)
-    let revert_bytecode = vec![0x60, 0x00, 0x60, 0x00, 0xFD];
-
-    let config = astar_runtime::RuntimeGenesisConfig {
-        system: Default::default(),
-        sudo: astar_runtime::SudoConfig {
-            key: Some(root_key),
-        },
-        parachain_info: ParachainInfoConfig {
-            parachain_id,
-            ..Default::default()
-        },
-        balances: astar_runtime::BalancesConfig { balances },
-        vesting: astar_runtime::VestingConfig { vesting: vec![] },
-        session: astar_runtime::SessionConfig {
-            keys: authorities
-                .iter()
-                .map(|x| (x.0.clone(), x.0.clone(), session_keys(x.1.clone())))
-                .collect::<Vec<_>>(),
-        },
-        aura: astar_runtime::AuraConfig {
-            authorities: vec![],
-        },
-        aura_ext: Default::default(),
-        collator_selection: astar_runtime::CollatorSelectionConfig {
-            desired_candidates: 32,
-            candidacy_bond: 3_200_000 * ASTR,
-            invulnerables: authorities.iter().map(|x| x.0.clone()).collect::<Vec<_>>(),
-        },
-        evm: EVMConfig {
-            // We need _some_ code inserted at the precompile address so that
-            // the evm will actually call the address.
-            accounts: Precompiles::used_addresses_h160()
-                .map(|addr| {
-                    (
-                        addr,
-                        fp_evm::GenesisAccount {
-                            nonce: Default::default(),
-                            balance: Default::default(),
-                            storage: Default::default(),
-                            code: revert_bytecode.clone(),
-                        },
-                    )
-                })
-                .collect(),
-            ..Default::default()
-        },
-        ethereum: Default::default(),
-        polkadot_xcm: Default::default(),
-        assets: Default::default(),
-        parachain_system: Default::default(),
-        transaction_payment: Default::default(),
-        dapp_staking: DappStakingConfig {
-            reward_portion: vec![
-                Permill::from_percent(40),
-                Permill::from_percent(30),
-                Permill::from_percent(20),
-                Permill::from_percent(10),
-            ],
-            slot_distribution: vec![
-                Permill::from_percent(10),
-                Permill::from_percent(20),
-                Permill::from_percent(30),
-                Permill::from_percent(40),
-            ],
-            // percentages below are calulated based on total issuance at the time when dApp staking v3 was launched (8.4B)
-            tier_thresholds: vec![
-                TierThreshold::DynamicPercentage {
-                    percentage: Perbill::from_parts(35_700_000), // 3.57%
-                    minimum_required_percentage: Perbill::from_parts(23_800_000), // 2.38%
-                },
-                TierThreshold::DynamicPercentage {
-                    percentage: Perbill::from_parts(8_900_000), // 0.89%
-                    minimum_required_percentage: Perbill::from_parts(6_000_000), // 0.6%
-                },
-                TierThreshold::DynamicPercentage {
-                    percentage: Perbill::from_parts(2_380_000), // 0.238%
-                    minimum_required_percentage: Perbill::from_parts(1_790_000), // 0.179%
-                },
-                TierThreshold::FixedPercentage {
-                    required_percentage: Perbill::from_parts(200_000), // 0.02%
-                },
-            ],
-            slots_per_tier: vec![10, 20, 30, 40],
-            safeguard: Some(false),
-            ..Default::default()
-        },
-        inflation: InflationConfig {
-            params: InflationParameters::default(),
-            ..Default::default()
-        },
-        oracle_membership: OracleMembershipConfig {
-            members: vec![
-                get_account_id_from_seed::<sr25519::Public>("Alice"),
-                get_account_id_from_seed::<sr25519::Public>("Bob"),
-            ]
-            .try_into()
-            .expect("Assumption is that at least two members will be allowed."),
-            ..Default::default()
-        },
-        price_aggregator: PriceAggregatorConfig {
-            circular_buffer: vec![CurrencyAmount::from_rational(5, 10)]
-                .try_into()
-                .expect("Must work since buffer should have at least a single value."),
-        },
-    };
-
-    serde_json::to_value(&config).expect("Could not build genesis config.")
-}
-
-type AccountPublic = <Signature as Verify>::Signer;
-
-/// Helper function to generate an account ID from seed
-fn get_account_id_from_seed<TPublic: Public>(seed: &str) -> AccountId
-where
-    AccountPublic: From<<TPublic::Pair as Pair>::Public>,
-{
-    AccountPublic::from(get_from_seed::<TPublic>(seed)).into_account()
 }
